Update base.rb to return multiple directors
The director method already returns an array, but if a movie (such as "The Big Lebowski" or "The Matrix") has more than one director, the second element in the returned array is "(more)". To fix this, I simply searched for directors in the fullcredits_document using the same format as the writers method.