Update previously changed 'directors' function
I changed the method name from 'directors' back to 'director'. It seems that the original 'director' method already returned the names of multiple directors and passed its spec test for 'Matrix Revolutions', but fails other movies with multiple directors such as 'The Big Lebowski' and 'Superman II'. This seems to be due to an inconsistency in the way that IMDB displays this data in the 'combined' view, from which @document gets its data. My changes to the director method pull director info from the @fullcredits_document which fixes the issue.