Add test for multiple director edge cases
Some movies with multiple directors (Superman II, Big Lebowski) display the second director's name as '(more)' in the combined view, and thus we need to pull director info from the fullcredits view to consistently return the correct results. I added a test on line 164 to test one of these cases.